Made it just do commonjs stuff well
