+++ bar.js
@@ -1,3 +1,2 @@
-var vars = require('./vars');
 
-module.exports = vars.createClass('Bar', ['getBarsAtEvent']);
+module.exports = require('./createClass')('Bar', ['getBarsAtEvent']);
