--- bower.json
@@ -1,29 +1,0 @@
-{
-  "name": "react-chartjs",
-  "main": "react-chartjs.js",
-  "version": "0.2.1",
-  "homepage": "https://github.com/jhudson8/react-chartjs",
-  "authors": [
-    "Joe Hudson <joehud@gmail.com>"
-  ],
-  "description": "rich interactive react charting components using chart.js",
-  "dependencies": {
-    "Chart.js": "~1.0"
-  },
-  "keywords": [
-    "react",
-    "react-component",
-    "chart",
-    "charts",
-    "graph",
-    "chartjs"
-  ],
-  "license": "MIT",
-  "ignore": [
-    "**/.*",
-    "node_modules",
-    "bower_components",
-    "test",
-    "tests"
-  ]
-}
