+++ doughnut.js
@@ -1,3 +1,2 @@
-var vars = require('./vars');
 
-module.exports = vars.createClass('Doughnut', ['getSegmentsAtEvent']);
+module.exports = require('./createClass')('Doughnut', ['getSegmentsAtEvent']);
