+++ line.js
@@ -1,3 +1,2 @@
-var vars = require('./vars');
 
-module.exports = vars.createClass('Line', ['getPointsAtEvent']);
+module.exports = require('./createClass')('Line', ['getPointsAtEvent']);
