+++ release-notes.md
@@ -1,25 +1,5 @@
 # Release Notes
 
-## v0.2.1 - January 26th, 2015
-- AMD bug fix
-
-
-## v0.2.0 - January 26th, 2015
-- Allow for commonJS code to not require initialization
-
-
-## v0.1.3 - January 19th, 2015
-- Actually include the commit mentioned with the 1.2 release
-
-
-## v0.1.2 - January 13th, 2015
-- fix chart rendering error when properties change
-
-
-## v0.1.1 - December 28th, 2014
-- fix AMD loader - 8c6b8ea
-
-
-## v0.1.0 - December 28th, 2014
-- add AMD loader - c04a10a
+## v0.0.2 - January 27th, 2015
+- Forked react-chartjs, and changed to only really support CommonJS, and do it seamlessly
