--- webpack.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-    entry: "./index.js",
-    output: {
-        path: __dirname,
-        filename: "react-chartjs.js"
-    }
-};
